+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 ### Fixed
 
 * Misspelt variable name in [HexaPDF::Layout::TextLayouter::SimpleLineWrapping]
+* [HexaPDF::Layout::TextLayouter::SimpleTextSegmentation] to work if the last
+  character in a text fragment is \r
 
 
 ## 0.6.0 - 2017-10-27